requirements.txt
hid==1.0.9
pillow==10.2.0
pystray==0.19.4
